Move local eslint rule to common
