+++ frontend/eslint-local-rules/index.js
@@ -1,5 +1,7 @@
-const matchingTranslationKeys = require('./matchingTranslationKeys.js')
+const path = require('path')
+const utils = require('eslint-plugin-vue/lib/utils/index.js')
+const matchingTranslationKeys = require('../src/common/eslint-local-rules/matchingTranslationKeys.js')
 
 module.exports = {
-  'matching-translation-keys': matchingTranslationKeys,
+  'matching-translation-keys': matchingTranslationKeys(path, utils),
 }
